requirements.txt
# Oak - Self-Improving AI Coding Agent
# Core (use >= to allow wheels compatible with Python 3.14)
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-multipart>=0.0.18
pydantic>=2.10.0
pydantic-settings>=2.7.0

# LLM
ollama>=0.4.0
httpx>=0.28.0

# Vector DB & Embeddings
chromadb>=0.6.0
sentence-transformers>=3.3.0

# GitHub Integration
PyGithub>=2.5.0
gitpython>=3.1.40

# Wiki / Markdown
markdown>=3.7
python-frontmatter>=1.1.0
beautifulsoup4>=4.12.0

# Speech-to-Text
faster-whisper>=1.0.0

# Utilities
psutil>=6.1.0
watchdog>=6.0.0
pygments>=2.18.0
aiofiles>=24.1.0
